+++ resource/wrapper/metricsresource/crud_resource_ops_wrapper_test.go
@@ -8,7 +8,7 @@ package metricsresource
 //	"testing"
 //
 //	"github.com/giantswarm/operatorkit/controller"
-//	"github.com/giantswarm/operatorkit/resource/internal"
+//	"github.com/giantswarm/operatorkit/resource/wrapper/internal"
 //)
 //
 //func Test_Wrapper(t *testing.T) {
