+++ resource/wrapper/metricsresource/crud_resource_wrapper.go
@@ -6,7 +6,7 @@ import (
 	"github.com/giantswarm/microerror"
 
 	"github.com/giantswarm/operatorkit/controller"
-	"github.com/giantswarm/operatorkit/resource/internal"
+	"github.com/giantswarm/operatorkit/resource/wrapper/internal"
 )
 
 // crudResourceWrapper is a specialized wrapper which wraps
