+++ resource/wrapper/retryresource/crud_resource_wrapper.go
@@ -9,7 +9,7 @@ import (
 	"github.com/giantswarm/micrologger"
 
 	"github.com/giantswarm/operatorkit/controller"
-	"github.com/giantswarm/operatorkit/resource/internal"
+	"github.com/giantswarm/operatorkit/resource/wrapper/internal"
 )
 
 // crudResourceWrapper is a specialized wrapper which wraps
